+++ src/main/java/com/github/mjkuranda/spaceadventure2/GameData.java
@@ -137,10 +137,11 @@ public class GameData {
         if (e != null) {
             player.damage(e.getDurability());
             spawn(new Particle(GameAnimation.EXPLOSION, e));
-            e.destroy();
+            e.remove();
 
             if (!player.isAlive()) {
-                game.enterState(StatesId.GAME_OVER);
+                game.enterState(StatesId.GAME_OVER_MENU);
+                in.clearKeyPressedRecord();
                 reset();
             }
         }