+++ src/main/java/com/github/mjkuranda/spaceadventure2/SpaceAdventure2.java
@@ -3,9 +3,9 @@ package com.github.mjkuranda.spaceadventure2;
 import com.github.mjkuranda.spaceadventure2.resources.GameAnimation;
 import com.github.mjkuranda.spaceadventure2.resources.GameFont;
 import com.github.mjkuranda.spaceadventure2.resources.GameImage;
-import com.github.mjkuranda.spaceadventure2.states.GameOverState;
 import com.github.mjkuranda.spaceadventure2.states.GameState;
 import com.github.mjkuranda.spaceadventure2.states.IntroState;
+import com.github.mjkuranda.spaceadventure2.states.menus.GameOverMenuState;
 import com.github.mjkuranda.spaceadventure2.states.menus.MainMenuState;
 import org.newdawn.slick.*;
 import org.newdawn.slick.state.StateBasedGame;
@@ -18,7 +18,7 @@ import java.util.logging.Logger;
 public class SpaceAdventure2 extends StateBasedGame {
 
     public static final String GAME_TITLE   = "Space Adventure II";
-    public static final String GAME_VERSION = "v0.14.0";
+    public static final String GAME_VERSION = "v0.15.0";
     public static final String GAME_CREATOR = "Marek Kuranda";
 
     public SpaceAdventure2() {
@@ -38,7 +38,7 @@ public class SpaceAdventure2 extends StateBasedGame {
 
         addState(new MainMenuState(this));
         addState(new GameState());
-        addState(new GameOverState());
+        addState(new GameOverMenuState(this));
         addState(new IntroState());
     }