+++ src/main/java/com/github/mjkuranda/spaceadventure2/states/menus/options/events/EnterStateMenuOptionEvent.java
@@ -1,5 +1,7 @@
 package com.github.mjkuranda.spaceadventure2.states.menus.options.events;
 
+import com.github.mjkuranda.spaceadventure2.PlayerStatistics;
+import org.newdawn.slick.Input;
 import org.newdawn.slick.state.StateBasedGame;
 
 public class EnterStateMenuOptionEvent implements MenuOptionEvent {
@@ -15,8 +17,12 @@ public class EnterStateMenuOptionEvent implements MenuOptionEvent {
     @Override
     public void onSelect() {
         game.enterState(stateId);
+        PlayerStatistics.getInstance().clear();
     }
 
     @Override
     public void onChange(int direction) {}
+
+    @Override
+    public void onChange(Input input) {}
 }