+++ src/main/java/com/github/mjkuranda/spaceadventure2/states/menus/options/events/MenuOptionEvent.java
@@ -1,5 +1,7 @@
 package com.github.mjkuranda.spaceadventure2.states.menus.options.events;
 
+import org.newdawn.slick.Input;
+
 public interface MenuOptionEvent {
     /***
      * Executes a specific task for menu option
@@ -11,4 +13,10 @@ public interface MenuOptionEvent {
      * @param direction -1 left, 1 right
      */
     void onChange(int direction);
+
+    /***
+     * Changes current value
+     * @param input org.newdawn.slick.Input
+     */
+    void onChange(Input input);
 }